feat: add review comments and task format
