+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_001.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 1",
+    "uuid": "f2e0456b-4870-5852-a56a-b924d8cb60d4",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the primary purpose of a CIF (Crystallographic Information File) in materials science?...",
+    "examples": [
+      {
+        "input": "What is the primary purpose of a CIF (Crystallographic Information File) in materials science?\n\nOptions:\n1. To store atomic coordinates and crystallographic data for a material structure.\n2. To simulate material properties using density functional theory.\n3. To visualize electron density distributions.\n4. To calculate the energy of a crystal lattice.",
+        "target_scores": {
+          "To store atomic coordinates and crystallographic data for a material structure.": 1,
+          "To simulate material properties using density functional theory.": 0,
+          "To visualize electron density distributions.": 0,
+          "To calculate the energy of a crystal lattice.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
