+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_002.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 2",
+    "uuid": "d8d0c504-dc88-547d-8655-39c31667ea9a",
+    "description": "Multiple choice question about Catalyst Environment QA: In the context of crystallography and surface science, what does the term 'Miller Index' represent?...",
+    "examples": [
+      {
+        "input": "In the context of crystallography and surface science, what does the term 'Miller Index' represent?\n\nOptions:\n1. A set of integers used to denote crystallographic planes and directions.\n2. The number of atoms in a unit cell.\n3. The lattice constant of a crystal.\n4. The adsorption energy of a molecule on a surface.",
+        "target_scores": {
+          "A set of integers used to denote crystallographic planes and directions.": 1,
+          "The number of atoms in a unit cell.": 0,
+          "The lattice constant of a crystal.": 0,
+          "The adsorption energy of a molecule on a surface.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
