+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_003.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 3",
+    "uuid": "10d0733d-bc09-51d1-9e01-86a70882157d",
+    "description": "Multiple choice question about Catalyst Environment QA: When constructing a slab model for surface simulations, why is it critical to appropriately define b...",
+    "examples": [
+      {
+        "input": "When constructing a slab model for surface simulations, why is it critical to appropriately define both the 'minimum slab size' and 'minimum vacuum size'?\n\nOptions:\n1. To directly determine the most energetically stable surface termination and reconstruction without further calculation.\n2. To ensure that the interior layers of the slab accurately mimic the bulk material's properties, uninfluenced by surface effects.\n3. To prevent unphysical electrostatic or van der Waals interactions between periodically repeated images of the slab across the simulation cell.\n4. To minimize the total number of atoms required for the simulation, thereby reducing computational demand significantly.\n5. To define the precise crystallographic orientation of the surface plane being studied.",
+        "target_scores": {
+          "To directly determine the most energetically stable surface termination and reconstruction without further calculation.": 0,
+          "To ensure that the interior layers of the slab accurately mimic the bulk material's properties, uninfluenced by surface effects.": 1,
+          "To prevent unphysical electrostatic or van der Waals interactions between periodically repeated images of the slab across the simulation cell.": 1,
+          "To minimize the total number of atoms required for the simulation, thereby reducing computational demand significantly.": 0,
+          "To define the precise crystallographic orientation of the surface plane being studied.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
