+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_004.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 4",
+    "uuid": "b9929d18-41ed-50c7-997d-8b04c216369a",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the purpose of enumerating possible slabs from a bulk structure?...",
+    "examples": [
+      {
+        "input": "What is the purpose of enumerating possible slabs from a bulk structure?\n\nOptions:\n1. To determine the bulk material's electronic properties.\n2. To create a 3D model of the bulk crystal.\n3. To explore different surface terminations and orientations.\n4. To find the most stable bulk crystal structure.",
+        "target_scores": {
+          "To determine the bulk material's electronic properties.": 0,
+          "To create a 3D model of the bulk crystal.": 0,
+          "To explore different surface terminations and orientations.": 1,
+          "To find the most stable bulk crystal structure.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
