+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_005.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 5",
+    "uuid": "820a405a-1686-5f41-b799-90d9d98e63f0",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the coordination number of titanium in TiO2?...",
+    "examples": [
+      {
+        "input": "What is the coordination number of titanium in TiO2?\n\nOptions:\n1. 4\n2. 6\n3. 8\n4. 12",
+        "target_scores": {
+          "4": 0,
+          "6": 1,
+          "8": 0,
+          "12": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
