+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_006.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 6",
+    "uuid": "e1c0129d-f72c-5579-a6b4-c0e055cacd38",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of the following describes the 'vacuum size' in a slab model used for surface simulations?...",
+    "examples": [
+      {
+        "input": "Which of the following describes the 'vacuum size' in a slab model used for surface simulations?\n\nOptions:\n1. The thickness of the slab itself.\n2. The volume occupied by the adsorbate molecule.\n3. The distance between the top surface of one slab and the bottom surface of its periodic image.\n4. The distance between atoms within the slab.",
+        "target_scores": {
+          "The thickness of the slab itself.": 0,
+          "The volume occupied by the adsorbate molecule.": 0,
+          "The distance between the top surface of one slab and the bottom surface of its periodic image.": 1,
+          "The distance between atoms within the slab.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
