+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_007.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 7",
+    "uuid": "ab5cca04-8842-5407-ac31-23bfb97e5ae4",
+    "description": "Multiple choice question about Catalyst Environment QA: Which TiO2 polymorph is thermodynamically most stable?...",
+    "examples": [
+      {
+        "input": "Which TiO2 polymorph is thermodynamically most stable?\n\nOptions:\n1. Anatase\n2. Rutile\n3. Brookite\n4. TiO2-B",
+        "target_scores": {
+          "Anatase": 0,
+          "Rutile": 1,
+          "Brookite": 0,
+          "TiO2-B": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
