+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_008.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 8",
+    "uuid": "c23fe2c6-cca3-5f3f-a0db-614f80c4b571",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the primary advantage of using a database like Materials Project (MP) for retrieving materia...",
+    "examples": [
+      {
+        "input": "What is the primary advantage of using a database like Materials Project (MP) for retrieving material structures?\n\nOptions:\n1. It can perform real-time simulations of material properties.\n2. It designs new materials based on user-defined criteria.\n3. It provides computationally validated materials data.\n4. Databases like NOMAD of OQMD also provide similar features",
+        "target_scores": {
+          "It can perform real-time simulations of material properties.": 0,
+          "It designs new materials based on user-defined criteria.": 0,
+          "It provides computationally validated materials data.": 1,
+          "Databases like NOMAD of OQMD also provide similar features": 1
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
