+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_009.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 9",
+    "uuid": "143dd28b-6206-5814-b8a5-eaa144f5d792",
+    "description": "Multiple choice question about Catalyst Environment QA: Which type of chemical bond is typically involved in the adsorption of a molecule onto a metal surfa...",
+    "examples": [
+      {
+        "input": "Which type of chemical bond is typically involved in the adsorption of a molecule onto a metal surface during chemisorption?\n\nOptions:\n1. Covalent bonds.\n2. Ionic bonds.\n3. Van der Waals forces.\n4. Hydrogen bonds.",
+        "target_scores": {
+          "Covalent bonds.": 1,
+          "Ionic bonds.": 0,
+          "Van der Waals forces.": 0,
+          "Hydrogen bonds.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
