+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_010.json
@@ -1,0 +1,35 @@
+[
+  {
+    "name": "Catalyst Environment QA 10",
+    "uuid": "ad615fee-2b83-5243-9b72-af5fe169a39e",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the role of Density Functional Theory (DFT) in computational materials science?...",
+    "examples": [
+      {
+        "input": "What is the role of Density Functional Theory (DFT) in computational materials science?\n\nOptions:\n1. To calculate the electronic structure of materials from first principles.\n2. To calculate the properties of materials from first principles.\n3. To optimize experimental crystal growth parameters in a laboratory setting.\n4. To approximate the many-body Schr\u00f6dinger equation by focusing on the electron density rather than the many-electron wave function.\n5. To apply Fourier transforms to directly measure experimental band gaps\n6. To solve the time-dependent Schr\u00f6dinger equation for simulating dynamics of many-body systems.\n7. To primarily study the mechanical properties of macroscopic bulk materials through finite element analysis.",
+        "target_scores": {
+          "To calculate the electronic structure of materials from first principles.": 1,
+          "To calculate the properties of materials from first principles.": 1,
+          "To optimize experimental crystal growth parameters in a laboratory setting.": 0,
+          "To approximate the many-body Schr\u00f6dinger equation by focusing on the electron density rather than the many-electron wave function.": 1,
+          "To apply Fourier transforms to directly measure experimental band gaps": 0,
+          "To solve the time-dependent Schr\u00f6dinger equation for simulating dynamics of many-body systems.": 0,
+          "To primarily study the mechanical properties of macroscopic bulk materials through finite element analysis.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
