+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_011.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 11",
+    "uuid": "b1bcaeb9-ee42-528e-8f9b-40259981c81d",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of the following statements correctly differentiates between an 'adsorbate' and an 'adsorbent'...",
+    "examples": [
+      {
+        "input": "Which of the following statements correctly differentiates between an 'adsorbate' and an 'adsorbent' in surface chemistry?\n\nOptions:\n1. The adsorbate is the surface itself, while the adsorbent is the substance that attaches to it.\n2. An adsorbate is the substance that adheres to a surface, whereas an adsorbent is the solid or liquid material providing that surface.\n3. Both adsorbate and adsorbent refer to the same chemical species, but one is in a gas phase and the other in a liquid phase.\n4. The adsorbent is a catalyst that facilitates adsorption, and the adsorbate is the product of that catalytic reaction.\n5. Adsorbate describes a substance dissolved in a solution, while adsorbent describes the solvent.",
+        "target_scores": {
+          "The adsorbate is the surface itself, while the adsorbent is the substance that attaches to it.": 0,
+          "An adsorbate is the substance that adheres to a surface, whereas an adsorbent is the solid or liquid material providing that surface.": 1,
+          "Both adsorbate and adsorbent refer to the same chemical species, but one is in a gas phase and the other in a liquid phase.": 0,
+          "The adsorbent is a catalyst that facilitates adsorption, and the adsorbate is the product of that catalytic reaction.": 0,
+          "Adsorbate describes a substance dissolved in a solution, while adsorbent describes the solvent.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
