+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_012.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 12",
+    "uuid": "093f3e0e-b4de-5342-85b5-a7e19dbd7b8b",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the crystal system of silicon?...",
+    "examples": [
+      {
+        "input": "What is the crystal system of silicon?\n\nOptions:\n1. Face-centered cubic\n2. Diamond cubic\n3. Simple cubic\n4. Body-centered cubic",
+        "target_scores": {
+          "Face-centered cubic": 0,
+          "Diamond cubic": 1,
+          "Simple cubic": 0,
+          "Body-centered cubic": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
