+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_013.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 13",
+    "uuid": "a7ae2163-ea3f-58b8-bbe2-e4a054aa8e72",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of the following statements about Cu2O (Cuprous Oxide) are generally true?...",
+    "examples": [
+      {
+        "input": "Which of the following statements about Cu2O (Cuprous Oxide) are generally true?\n\nOptions:\n1. It is a metallic conductor.\n2. It is commonly used as a transparent conductor.\n3. It is a p-type semiconductor.\n4. It is primarily used in high-temperature superconducting applications.",
+        "target_scores": {
+          "It is a metallic conductor.": 0,
+          "It is commonly used as a transparent conductor.": 0,
+          "It is a p-type semiconductor.": 1,
+          "It is primarily used in high-temperature superconducting applications.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
