+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_014.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 14",
+    "uuid": "99f8779a-0370-5c4e-8d84-68d8059b2875",
+    "description": "Multiple choice question about Catalyst Environment QA: Why is the (111) surface often a common choice for initial studies in surface science?...",
+    "examples": [
+      {
+        "input": "Why is the (111) surface often a common choice for initial studies in surface science?\n\nOptions:\n1. It is frequently one of the most stable or densely packed surfaces for many crystal structures.\n2. It always has the highest surface energy.\n3. It is exclusively found in body-centered cubic (BCC) crystals.\n4. It is the only surface that exhibits catalytic activity.",
+        "target_scores": {
+          "It is frequently one of the most stable or densely packed surfaces for many crystal structures.": 1,
+          "It always has the highest surface energy.": 0,
+          "It is exclusively found in body-centered cubic (BCC) crystals.": 0,
+          "It is the only surface that exhibits catalytic activity.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
