+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_015.json
@@ -1,0 +1,38 @@
+[
+  {
+    "name": "Catalyst Environment QA 15",
+    "uuid": "dee1e66f-13d6-5c87-85a4-29eafa0fd6a8",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of these factors can significantly influence the adsorption site and the energy (strength) of ...",
+    "examples": [
+      {
+        "input": "Which of these factors can significantly influence the adsorption site and the energy (strength) of interaction for a specific molecule on a material's surface?\n\nOptions:\n1. The atomic arrangement and coordination environment of the surface atoms.\n2. The chemical identity, molecular geometry, and electronic configuration of the adsorbate molecule.\n3. The work function of the bulk material.\n4. The electronic band structure and charge density distribution at the surface.\n5. The bulk modulus of the underlying material.\n6. The melting point of the adsorbent.\n7. The presence and concentration of pre-adsorbed species on the surface.\n8. The temperature and pressure conditions during the adsorption process.\n9. The diffusion coefficient of the adsorbate within the bulk material.\n10. The specific heat capacity of the adsorbent.",
+        "target_scores": {
+          "The atomic arrangement and coordination environment of the surface atoms.": 1,
+          "The chemical identity, molecular geometry, and electronic configuration of the adsorbate molecule.": 1,
+          "The work function of the bulk material.": 0,
+          "The electronic band structure and charge density distribution at the surface.": 1,
+          "The bulk modulus of the underlying material.": 0,
+          "The melting point of the adsorbent.": 0,
+          "The presence and concentration of pre-adsorbed species on the surface.": 1,
+          "The temperature and pressure conditions during the adsorption process.": 1,
+          "The diffusion coefficient of the adsorbate within the bulk material.": 0,
+          "The specific heat capacity of the adsorbent.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
