+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_016.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 16",
+    "uuid": "b793069a-6412-5f3f-ab7f-1bcf3cbef8a1",
+    "description": "Multiple choice question about Catalyst Environment QA: What happens when a molecule undergoes physisorption on a surface?...",
+    "examples": [
+      {
+        "input": "What happens when a molecule undergoes physisorption on a surface?\n\nOptions:\n1. Weak, non-covalent interactions (e.g., Van der Waals forces) hold the molecule to the surface.\n2. A strong chemical bond is formed between the molecule and the surface.\n3. The molecule undergoes a chemical reaction with the surface.\n4. Electron transfer occurs, leading to ionization of the molecule.",
+        "target_scores": {
+          "Weak, non-covalent interactions (e.g., Van der Waals forces) hold the molecule to the surface.": 1,
+          "A strong chemical bond is formed between the molecule and the surface.": 0,
+          "The molecule undergoes a chemical reaction with the surface.": 0,
+          "Electron transfer occurs, leading to ionization of the molecule.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
