+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_017.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 17",
+    "uuid": "4f112fd4-8534-591b-8256-83b004a166b8",
+    "description": "Multiple choice question about Catalyst Environment QA: Which property makes TiO2 particularly useful for photocatalysis?...",
+    "examples": [
+      {
+        "input": "Which property makes TiO2 particularly useful for photocatalysis?\n\nOptions:\n1. High refractive index\n2. Chemical stability\n3. UV absorption\n4. Low cost",
+        "target_scores": {
+          "High refractive index": 0,
+          "Chemical stability": 1,
+          "UV absorption": 1,
+          "Low cost": 1
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
