+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_018.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 18",
+    "uuid": "b5eb1d6e-fe2b-59f1-8e28-6f6729b9b1b9",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the 'coordination number' of an atom in a crystal structure?...",
+    "examples": [
+      {
+        "input": "What is the 'coordination number' of an atom in a crystal structure?\n\nOptions:\n1. The atomic weight of the element.\n2. The number of valence electrons.\n3. The density of the material.\n4. The number of neighbor atoms to a central atom which is bonded to the central atom.\n5. The number of nearest neighbor atoms to a central atom.",
+        "target_scores": {
+          "The atomic weight of the element.": 0,
+          "The number of valence electrons.": 0,
+          "The density of the material.": 0,
+          "The number of neighbor atoms to a central atom which is bonded to the central atom.": 0,
+          "The number of nearest neighbor atoms to a central atom.": 1
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
