+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_019.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 19",
+    "uuid": "7f4b9222-7f6e-5524-ab25-db45bd6d3507",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of the following can be considered types of adsorption sites on a crystalline surface?...",
+    "examples": [
+      {
+        "input": "Which of the following can be considered types of adsorption sites on a crystalline surface?\n\nOptions:\n1. Ontop site\n2. Bridge site\n3. Hollow site\n4. Interstitial site\n5. Schottky site",
+        "target_scores": {
+          "Ontop site": 1,
+          "Bridge site": 1,
+          "Hollow site": 1,
+          "Interstitial site": 0,
+          "Schottky site": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
