+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_020.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 20",
+    "uuid": "05d1f0d3-441e-53b9-be29-e84464e38276",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the primary reason for performing geometry optimization after placing an adsorbate on a slab...",
+    "examples": [
+      {
+        "input": "What is the primary reason for performing geometry optimization after placing an adsorbate on a slab?\n\nOptions:\n1. To find the lowest energy (most stable) configuration of the system.\n2. To increase the temperature of the system.\n3. To introduce defects into the crystal structure.\n4. To calculate the electronic band structure.",
+        "target_scores": {
+          "To find the lowest energy (most stable) configuration of the system.": 1,
+          "To increase the temperature of the system.": 0,
+          "To introduce defects into the crystal structure.": 0,
+          "To calculate the electronic band structure.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
