+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_021.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 21",
+    "uuid": "fc7ecb24-ef10-5ec7-8471-8e422cc0e333",
+    "description": "Multiple choice question about Catalyst Environment QA: What type of semiconductor is pure silicon?...",
+    "examples": [
+      {
+        "input": "What type of semiconductor is pure silicon?\n\nOptions:\n1. Metal\n2. Intrinsic semiconductor\n3. n-type semiconductor\n4. p-type semiconductor",
+        "target_scores": {
+          "Metal": 0,
+          "Intrinsic semiconductor": 1,
+          "n-type semiconductor": 0,
+          "p-type semiconductor": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
