+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_022.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 22",
+    "uuid": "ac429d47-da6b-50d9-9ae6-c23992664404",
+    "description": "Multiple choice question about Catalyst Environment QA: Why is it important to consider the charge state of atoms (e.g., Cu+ in Cu2O) when analyzing bonding...",
+    "examples": [
+      {
+        "input": "Why is it important to consider the charge state of atoms (e.g., Cu+ in Cu2O) when analyzing bonding and reactivity?\n\nOptions:\n1. The charge state is only relevant for metallic materials.\n2. The charge state influences electron distribution, which affects chemical bonding and interactions.\n3. The charge state directly determines the material's color.\n4. The charge state dictates the material's melting point.",
+        "target_scores": {
+          "The charge state is only relevant for metallic materials.": 0,
+          "The charge state influences electron distribution, which affects chemical bonding and interactions.": 1,
+          "The charge state directly determines the material's color.": 0,
+          "The charge state dictates the material's melting point.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
