+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_023.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 23",
+    "uuid": "317507bd-9009-5c62-8986-99a5a563a3c4",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of these are common challenges when building accurate slab models for surface simulations?...",
+    "examples": [
+      {
+        "input": "Which of these are common challenges when building accurate slab models for surface simulations?\n\nOptions:\n1. Determining the correct surface termination (e.g., Cu-terminated vs. O-terminated).\n2. Computing the correct bulk properties of the material.\n3. Ensuring sufficient slab thickness to avoid artificial interactions between surfaces.\n4. Choosing an appropriate vacuum layer thickness.\n5. Obtaining the correct bulk modulus of the material.",
+        "target_scores": {
+          "Determining the correct surface termination (e.g., Cu-terminated vs. O-terminated).": 1,
+          "Computing the correct bulk properties of the material.": 0,
+          "Ensuring sufficient slab thickness to avoid artificial interactions between surfaces.": 1,
+          "Choosing an appropriate vacuum layer thickness.": 1,
+          "Obtaining the correct bulk modulus of the material.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
