+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_024.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 24",
+    "uuid": "f0716f11-c29d-5250-8460-8a21a7eb2abd",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of these tools or techniques are commonly used to characterize adsorption sites experimentally...",
+    "examples": [
+      {
+        "input": "Which of these tools or techniques are commonly used to characterize adsorption sites experimentally?\n\nOptions:\n1. Scanning Tunneling Microscopy (STM)\n2. Low-Energy Electron Diffraction (LEED)\n3. X-ray Diffraction (XRD) of bulk crystals\n4. UV-Vis Spectroscopy",
+        "target_scores": {
+          "Scanning Tunneling Microscopy (STM)": 1,
+          "Low-Energy Electron Diffraction (LEED)": 1,
+          "X-ray Diffraction (XRD) of bulk crystals": 0,
+          "UV-Vis Spectroscopy": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
