+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_025.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 25",
+    "uuid": "075a80e7-31ed-5b4a-8d7c-2778ffcf2d3e",
+    "description": "Multiple choice question about Catalyst Environment QA: In the context of surface energy, what generally happens to the atoms at the surface of a material c...",
+    "examples": [
+      {
+        "input": "In the context of surface energy, what generally happens to the atoms at the surface of a material compared to those in the bulk?\n\nOptions:\n1. They have lower energy due to increased atomic density.\n2. They are always perfectly aligned with bulk atoms.\n3. They have higher energy due to fewer nearest neighbors and unsaturated bonds.\n4. They undergo complete chemical transformation.",
+        "target_scores": {
+          "They have lower energy due to increased atomic density.": 0,
+          "They are always perfectly aligned with bulk atoms.": 0,
+          "They have higher energy due to fewer nearest neighbors and unsaturated bonds.": 1,
+          "They undergo complete chemical transformation.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
