+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_026.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 26",
+    "uuid": "26137f9e-2632-5b12-85e6-ef1bdbf98949",
+    "description": "Multiple choice question about Catalyst Environment QA: Which crystal structure does anatase TiO2 have?...",
+    "examples": [
+      {
+        "input": "Which crystal structure does anatase TiO2 have?\n\nOptions:\n1. Tetragonal\n2. Cubic\n3. Hexagonal\n4. Orthorhombic",
+        "target_scores": {
+          "Tetragonal": 1,
+          "Cubic": 0,
+          "Hexagonal": 0,
+          "Orthorhombic": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
