+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_027.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 27",
+    "uuid": "1dc5e7ca-c99d-58c9-8fc8-801268a6ead2",
+    "description": "Multiple choice question about Catalyst Environment QA: What is a 'unit cell' in crystallography?...",
+    "examples": [
+      {
+        "input": "What is a 'unit cell' in crystallography?\n\nOptions:\n1. The smallest unit of a crystal lattice that determines all the catalytic properties of that material\n2. The entire crystal structure.\n3. The smallest repeating unit that generates the entire crystal when translated in three dimensions.\n4. A single atom within the crystal.",
+        "target_scores": {
+          "The smallest unit of a crystal lattice that determines all the catalytic properties of that material": 0,
+          "The entire crystal structure.": 0,
+          "The smallest repeating unit that generates the entire crystal when translated in three dimensions.": 1,
+          "A single atom within the crystal.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
