+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_028.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 28",
+    "uuid": "eb59a8c7-94a2-51ce-95b7-e79ff06c6062",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the crystal structure of Cu2O (cuprous oxide)?...",
+    "examples": [
+      {
+        "input": "What is the crystal structure of Cu2O (cuprous oxide)?\n\nOptions:\n1. Face-centered cubic (FCC)\n2. Body-centered cubic (BCC)\n3. Simple cubic\n4. Hexagonal close-packed",
+        "target_scores": {
+          "Face-centered cubic (FCC)": 0,
+          "Body-centered cubic (BCC)": 0,
+          "Simple cubic": 1,
+          "Hexagonal close-packed": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
