+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_029.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 29",
+    "uuid": "c5eafee2-ee6b-5d77-89b2-f4258e927e86",
+    "description": "Multiple choice question about Catalyst Environment QA: Which Miller indices represent the most densely packed plane in an FCC crystal?...",
+    "examples": [
+      {
+        "input": "Which Miller indices represent the most densely packed plane in an FCC crystal?\n\nOptions:\n1. (100)\n2. (110)\n3. (111)\n4. (200)",
+        "target_scores": {
+          "(100)": 0,
+          "(110)": 0,
+          "(111)": 1,
+          "(200)": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
