+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_030.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 30",
+    "uuid": "d13e7f52-fb81-5df1-abd6-4332ca358d3b",
+    "description": "Multiple choice question about Catalyst Environment QA: What does a CIF file contain?...",
+    "examples": [
+      {
+        "input": "What does a CIF file contain?\n\nOptions:\n1. Computational parameters\n2. Only atomic coordinates\n3. Crystal structure information\n4. Electronic band structure",
+        "target_scores": {
+          "Computational parameters": 0,
+          "Only atomic coordinates": 0,
+          "Crystal structure information": 1,
+          "Electronic band structure": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
