+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_031.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 31",
+    "uuid": "897b8581-158c-534d-8aa3-bcd50a1a8392",
+    "description": "Multiple choice question about Catalyst Environment QA: In surface science, what is an 'ontop' adsorption site?...",
+    "examples": [
+      {
+        "input": "In surface science, what is an 'ontop' adsorption site?\n\nOptions:\n1. Between two surface atoms\n2. Directly above a surface atom\n3. In a hollow formed by three atoms\n4. Below the surface layer",
+        "target_scores": {
+          "Between two surface atoms": 0,
+          "Directly above a surface atom": 1,
+          "In a hollow formed by three atoms": 0,
+          "Below the surface layer": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
