+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_032.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 32",
+    "uuid": "0b12dc54-34f1-5275-88e4-3cc147d8293b",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the coordination number of copper in Cu2O?...",
+    "examples": [
+      {
+        "input": "What is the coordination number of copper in Cu2O?\n\nOptions:\n1. 2\n2. 4\n3. 6\n4. 8",
+        "target_scores": {
+          "2": 1,
+          "4": 0,
+          "6": 0,
+          "8": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
