+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_033.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 33",
+    "uuid": "8b8a105a-1c87-5960-b92c-9aecca60c5d4",
+    "description": "Multiple choice question about Catalyst Environment QA: When placing an adsorbate onto a slab surface, what is the primary role of specifying an initial 'he...",
+    "examples": [
+      {
+        "input": "When placing an adsorbate onto a slab surface, what is the primary role of specifying an initial 'height' parameter for the adsorbate molecule?\n\nOptions:\n1. It precisely determines the final, equilibrium adsorption energy of the system.\n2. It controls the lateral position and rotational orientation of the adsorbate on the surface.\n3. It establishes the starting vertical distance of the adsorbate from the surface for subsequent structural optimization.\n4. It is a fixed parameter that defines the thickness of the surface slab itself.\n5. It ensures that the adsorbate immediately forms the strongest possible chemical bond with the surface.",
+        "target_scores": {
+          "It precisely determines the final, equilibrium adsorption energy of the system.": 0,
+          "It controls the lateral position and rotational orientation of the adsorbate on the surface.": 0,
+          "It establishes the starting vertical distance of the adsorbate from the surface for subsequent structural optimization.": 1,
+          "It is a fixed parameter that defines the thickness of the surface slab itself.": 0,
+          "It ensures that the adsorbate immediately forms the strongest possible chemical bond with the surface.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
