+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_034.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 34",
+    "uuid": "c800a8c3-7ad6-5eee-a1a4-c5cd427b769d",
+    "description": "Multiple choice question about Catalyst Environment QA: Which factors affect CO2 adsorption on metal oxide surfaces?...",
+    "examples": [
+      {
+        "input": "Which factors affect CO2 adsorption on metal oxide surfaces?\n\nOptions:\n1. Surface defects\n2. Temperature\n3. Surface orientation\n4. Vacuum level",
+        "target_scores": {
+          "Surface defects": 1,
+          "Temperature": 1,
+          "Surface orientation": 1,
+          "Vacuum level": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
