+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_035.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 35",
+    "uuid": "f0ea2a50-dbbc-5467-aa28-0622828fb2c2",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the typical range for vacuum thickness in slab calculations?...",
+    "examples": [
+      {
+        "input": "What is the typical range for vacuum thickness in slab calculations?\n\nOptions:\n1. 1-2 \u00c5\n2. 10-15 \u00c5\n3. 50-100 \u00c5\n4. 200+ \u00c5",
+        "target_scores": {
+          "1-2 \u00c5": 0,
+          "10-15 \u00c5": 1,
+          "50-100 \u00c5": 0,
+          "200+ \u00c5": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
