+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_036.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 36",
+    "uuid": "da9490c4-c0b6-5124-903a-33dc58d86c69",
+    "description": "Multiple choice question about Catalyst Environment QA: What does the Miller index (111) represent?...",
+    "examples": [
+      {
+        "input": "What does the Miller index (111) represent?\n\nOptions:\n1. A plane that intercepts all three axes at unit distances\n2. A direction in the crystal\n3. The crystal's lattice parameter\n4. The number of atoms in the unit cell",
+        "target_scores": {
+          "A plane that intercepts all three axes at unit distances": 1,
+          "A direction in the crystal": 0,
+          "The crystal's lattice parameter": 0,
+          "The number of atoms in the unit cell": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
