+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_037.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 37",
+    "uuid": "54f9d605-645f-5282-bf17-73c0d1cd2393",
+    "description": "Multiple choice question about Catalyst Environment QA: What is surface reconstruction?...",
+    "examples": [
+      {
+        "input": "What is surface reconstruction?\n\nOptions:\n1. Rearrangement of surface atoms from bulk positions\n2. Addition of new atomic layers\n3. Removal of surface atoms\n4. Change in crystal structure",
+        "target_scores": {
+          "Rearrangement of surface atoms from bulk positions": 1,
+          "Addition of new atomic layers": 0,
+          "Removal of surface atoms": 0,
+          "Change in crystal structure": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
