+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_038.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 38",
+    "uuid": "6daf0a61-006c-5c0c-841d-f870346edaa7",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the coordination environment of oxygen in Cu2O?...",
+    "examples": [
+      {
+        "input": "What is the coordination environment of oxygen in Cu2O?\n\nOptions:\n1. Linear\n2. Trigonal planar\n3. Tetrahedral\n4. Octahedral",
+        "target_scores": {
+          "Linear": 0,
+          "Trigonal planar": 0,
+          "Tetrahedral": 1,
+          "Octahedral": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
