+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_039.json
@@ -1,0 +1,31 @@
+[
+  {
+    "name": "Catalyst Environment QA 39",
+    "uuid": "cd2e8426-ef11-5cf8-8e57-efbc31ec28a8",
+    "description": "Multiple choice question about Catalyst Environment QA: Which factors affect the minimum slab thickness in calculations?...",
+    "examples": [
+      {
+        "input": "Which factors affect the minimum slab thickness in calculations?\n\nOptions:\n1. Convergence of surface properties\n2. Computational cost\n3. Surface termination",
+        "target_scores": {
+          "Convergence of surface properties": 1,
+          "Computational cost": 1,
+          "Surface termination": 1
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
