+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_040.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 40",
+    "uuid": "1ca71256-930c-5c3d-80b0-35c81b4f2a7a",
+    "description": "Multiple choice question about Catalyst Environment QA: What is the relationship between surface energy and surface stability?...",
+    "examples": [
+      {
+        "input": "What is the relationship between surface energy and surface stability?\n\nOptions:\n1. Higher surface energy = more stable\n2. Lower surface energy = more stable\n3. No relationship\n4. Inversely proportional to temperature",
+        "target_scores": {
+          "Higher surface energy = more stable": 0,
+          "Lower surface energy = more stable": 1,
+          "No relationship": 0,
+          "Inversely proportional to temperature": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
