+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_041.json
@@ -1,0 +1,31 @@
+[
+  {
+    "name": "Catalyst Environment QA 41",
+    "uuid": "dca990b6-e7f1-51dc-a394-a04457f5d60c",
+    "description": "Multiple choice question about Catalyst Environment QA: Which defects are commonly found on TiO2 surfaces?...",
+    "examples": [
+      {
+        "input": "Which defects are commonly found on TiO2 surfaces?\n\nOptions:\n1. Oxygen vacancies\n2. Titanium interstitials\n3. Carbon substitutions",
+        "target_scores": {
+          "Oxygen vacancies": 1,
+          "Titanium interstitials": 1,
+          "Carbon substitutions": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
