+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_042.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 42",
+    "uuid": "b6956045-09a3-5e6b-aab3-8d08e0510aa5",
+    "description": "Multiple choice question about Catalyst Environment QA: Which dopants are commonly used to make silicon n-type?...",
+    "examples": [
+      {
+        "input": "Which dopants are commonly used to make silicon n-type?\n\nOptions:\n1. Phosphorus\n2. Boron\n3. Arsenic\n4. Aluminum",
+        "target_scores": {
+          "Phosphorus": 1,
+          "Boron": 0,
+          "Arsenic": 1,
+          "Aluminum": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
