+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_043.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 43",
+    "uuid": "27fef822-f1d8-560c-8807-ea19ce67977d",
+    "description": "Multiple choice question about Catalyst Environment QA: Which factors most influences TiO2 photocatalytic efficiency?...",
+    "examples": [
+      {
+        "input": "Which factors most influences TiO2 photocatalytic efficiency?\n\nOptions:\n1. Crystal size\n2. Surface area\n3. Defect concentration\n4. Bandgap",
+        "target_scores": {
+          "Crystal size": 1,
+          "Surface area": 1,
+          "Defect concentration": 1,
+          "Bandgap": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
