+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_044.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 44",
+    "uuid": "286014e2-0542-579a-bdd2-5725a98a0868",
+    "description": "Multiple choice question about Catalyst Environment QA: Which two DFT functionals are considered higher levels of theory among the following?...",
+    "examples": [
+      {
+        "input": "Which two DFT functionals are considered higher levels of theory among the following?\n\nOptions:\n1. Local Density Approximation (LDA)\n2. Generalized Gradient Approximation (GGA)\n3. Hybrid functionals (B3LYP, PBE0)\n4. Meta-GGA functionals\n5. Hartree-Fock",
+        "target_scores": {
+          "Local Density Approximation (LDA)": 0,
+          "Generalized Gradient Approximation (GGA)": 0,
+          "Hybrid functionals (B3LYP, PBE0)": 1,
+          "Meta-GGA functionals": 1,
+          "Hartree-Fock": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
