+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_045.json
@@ -1,0 +1,40 @@
+[
+  {
+    "name": "Catalyst Environment QA 45",
+    "uuid": "8efa930c-cc85-5cb2-8ddc-f8a88f3c5b29",
+    "description": "Multiple choice question about Catalyst Environment QA: Which of the following computational parameters or choices directly and fundamentally influence the ...",
+    "examples": [
+      {
+        "input": "Which of the following computational parameters or choices directly and fundamentally influence the *accuracy* of calculated molecular adsorption energies in periodic Density Functional Theory (DFT) simulations?\n\nOptions:\n1. The number of atomic layers included in the surface slab model.\n2. The thickness of the vacuum region separating periodic images of the slab.\n3. The density of the k-point mesh used for Brillouin zone integration.\n4. The specific form of the exchange-correlation functional employed.\n5. The total simulation time of a molecular dynamics trajectory.\n6. The computational power (CPU cores, RAM) of the hardware cluster.\n7. The visualization software used to render atomic structures.\n8. The completeness and quality of the chosen basis set.\n9. The convergence criterion for the self-consistent field (SCF) loop.\n10. The choice of pseudopotential type.\n11. The type of thermostat used in a molecular dynamics simulation.\n12. The compiler optimization flags used for the DFT code.",
+        "target_scores": {
+          "The number of atomic layers included in the surface slab model.": 1,
+          "The thickness of the vacuum region separating periodic images of the slab.": 1,
+          "The density of the k-point mesh used for Brillouin zone integration.": 1,
+          "The specific form of the exchange-correlation functional employed.": 1,
+          "The total simulation time of a molecular dynamics trajectory.": 0,
+          "The computational power (CPU cores, RAM) of the hardware cluster.": 0,
+          "The visualization software used to render atomic structures.": 0,
+          "The completeness and quality of the chosen basis set.": 1,
+          "The convergence criterion for the self-consistent field (SCF) loop.": 1,
+          "The choice of pseudopotential type.": 1,
+          "The type of thermostat used in a molecular dynamics simulation.": 0,
+          "The compiler optimization flags used for the DFT code.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
