+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_046.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "Catalyst Environment QA 46",
+    "uuid": "eb8e1b78-644a-571a-b26d-bd17d1ea4c1f",
+    "description": "Multiple choice question about Catalyst Environment QA: When simulating molecular adsorption on a crystalline surface using DFT, which of the following is a...",
+    "examples": [
+      {
+        "input": "When simulating molecular adsorption on a crystalline surface using DFT, which of the following is an important consideration or common assumption regarding the surface structure upon adsorption?\n\nOptions:\n1. The surface always undergoes significant reconstruction into a completely different symmetry.\n2. The substrate atoms are always rigidly fixed in their bulk termination positions.\n3. The surface atoms may relax or reconstruct locally in response to the adsorbate, which should be considered for accurate binding energies.\n4. The adsorbate only interacts with the topmost layer of surface atoms, ignoring deeper interactions.\n5. The adsorbate's internal molecular geometry always remains rigid during adsorption.",
+        "target_scores": {
+          "The surface always undergoes significant reconstruction into a completely different symmetry.": 0,
+          "The substrate atoms are always rigidly fixed in their bulk termination positions.": 0,
+          "The surface atoms may relax or reconstruct locally in response to the adsorbate, which should be considered for accurate binding energies.": 1,
+          "The adsorbate only interacts with the topmost layer of surface atoms, ignoring deeper interactions.": 0,
+          "The adsorbate's internal molecular geometry always remains rigid during adsorption.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
