+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_047.json
@@ -1,0 +1,36 @@
+[
+  {
+    "name": "Catalyst Environment QA 47",
+    "uuid": "70be71fd-3f9a-5e31-af79-94b74f96791b",
+    "description": "Multiple choice question about Catalyst Environment QA: What are the main advantages of plane wave basis sets in DFT calculations?...",
+    "examples": [
+      {
+        "input": "What are the main advantages of plane wave basis sets in DFT calculations?\n\nOptions:\n1. Systematic convergence with energy cutoff\n2. Efficient for periodic systems\n3. Good for localized orbitals\n4. Fast for molecules in vacuum\n5. Natural treatment of periodicity\n6. They require explicit treatment of core electrons, leading to higher computational cost.\n7. They intrinsically fulfill the boundary conditions for molecular clusters in vacuum.\n8. They are primarily used in classical force-field molecular dynamics simulations.",
+        "target_scores": {
+          "Systematic convergence with energy cutoff": 1,
+          "Efficient for periodic systems": 1,
+          "Good for localized orbitals": 0,
+          "Fast for molecules in vacuum": 0,
+          "Natural treatment of periodicity": 1,
+          "They require explicit treatment of core electrons, leading to higher computational cost.": 0,
+          "They intrinsically fulfill the boundary conditions for molecular clusters in vacuum.": 0,
+          "They are primarily used in classical force-field molecular dynamics simulations.": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
