+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_048.json
@@ -1,0 +1,35 @@
+[
+  {
+    "name": "Catalyst Environment QA 48",
+    "uuid": "6144c9f3-865b-5d57-9479-20ef5a2813a2",
+    "description": "Multiple choice question about Catalyst Environment QA: In what scenario are Gaussian-type orbital basis sets typically preferred over plane-wave basis sets...",
+    "examples": [
+      {
+        "input": "In what scenario are Gaussian-type orbital basis sets typically preferred over plane-wave basis sets in quantum mechanical simulations?\n\nOptions:\n1. When simulating bulk metallic systems with strong delocalization.\n2. For calculations on periodic crystalline solids where precise long-range interactions are critical.\n3. When performing high-accuracy calculations on isolated molecules or finite clusters in vacuum.\n4. Well-suited for molecular calculations\n5. Computationally efficient for isolated systems\n6. Always more accurate than plane waves\n7. Natural for periodic boundary conditions",
+        "target_scores": {
+          "When simulating bulk metallic systems with strong delocalization.": 0,
+          "For calculations on periodic crystalline solids where precise long-range interactions are critical.": 0,
+          "When performing high-accuracy calculations on isolated molecules or finite clusters in vacuum.": 1,
+          "Well-suited for molecular calculations": 1,
+          "Computationally efficient for isolated systems": 1,
+          "Always more accurate than plane waves": 0,
+          "Natural for periodic boundary conditions": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
