+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_049.json
@@ -1,0 +1,35 @@
+[
+  {
+    "name": "Catalyst Environment QA 49",
+    "uuid": "6042221c-c7b7-5098-b703-ff4ef7d4f70a",
+    "description": "Multiple choice question about Catalyst Environment QA: According to the Sabatier principle, what factors determine the optimal catalyst for CO2 reduction t...",
+    "examples": [
+      {
+        "input": "According to the Sabatier principle, what factors determine the optimal catalyst for CO2 reduction to CO, and how do scaling relations constrain catalyst design?\n\nOptions:\n1. d-band center position controls adsorption strength; optimal activity requires moderate CO binding; scaling relations between CO and CHO binding limit selectivity\n2. Only surface area matters for CO2 reduction; d-band theory doesn't apply to oxide catalysts\n3. Stronger CO binding always leads to better activity; scaling relations can be easily broken by changing particle size\n4. d-band center should be as close to Fermi level as possible; there are no fundamental limits to catalyst optimization\n5. Electronic effects are secondary to geometric effects; optimal catalysts always have the highest CO binding energy\n6. Volcano plots show linear relationships between binding energy and activity; scaling relations only affect transition metals\n7. Volcano plots demonstrate the Sabatier principle with peak activity at intermediate binding energies; the optimal d-band center corresponds to the volcano maximum",
+        "target_scores": {
+          "d-band center position controls adsorption strength; optimal activity requires moderate CO binding; scaling relations between CO and CHO binding limit selectivity": 1,
+          "Only surface area matters for CO2 reduction; d-band theory doesn't apply to oxide catalysts": 0,
+          "Stronger CO binding always leads to better activity; scaling relations can be easily broken by changing particle size": 0,
+          "d-band center should be as close to Fermi level as possible; there are no fundamental limits to catalyst optimization": 0,
+          "Electronic effects are secondary to geometric effects; optimal catalysts always have the highest CO binding energy": 0,
+          "Volcano plots show linear relationships between binding energy and activity; scaling relations only affect transition metals": 0,
+          "Volcano plots demonstrate the Sabatier principle with peak activity at intermediate binding energies; the optimal d-band center corresponds to the volcano maximum": 1
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
