+++ tasks/qa/catalyst_qa/tasks_json/catalyst_question_050.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Catalyst Environment QA 50",
+    "uuid": "478b8fde-1d84-5f95-a22a-9f7f1680fa32",
+    "description": "Multiple choice question about Catalyst Environment QA: You're calculating CO adsorption energies on different metal surfaces from DFT calculations. Which i...",
+    "examples": [
+      {
+        "input": "You're calculating CO adsorption energies on different metal surfaces from DFT calculations. Which implementation correctly calculates the adsorption energy?\n\n```python\ndef calculate_adsorption_energies(dft_results):\n    \"\"\"\n    Calculate CO adsorption energy on metal surfaces.\n    DFT results contain:\n    - 'clean_slab': Energy of clean metal slab\n    - 'co_adsorbate': Energy of CO molecule in gas phase\n    - 'slab_co': Energy of slab + adsorbed CO system\n    \"\"\"\n    adsorption_energies = {}\n    \n    for surface in ['Pt111', 'Pd111', 'Ni111']:\n        E_clean = dft_results[surface]['clean_slab']\n        E_co_gas = dft_results['CO']['co_adsorbate']\n        E_slab_co = dft_results[surface]['slab_co']\n        \n        # Different implementations:\n```\n\nOptions:\n1. ```python\n# Implementation A\nE_ads = E_slab_co - E_clean - E_co_gas\nadsorption_energies[surface] = E_ads\n```\n2. ```python\n# Implementation B\nE_ads = E_clean + E_co_gas - E_slab_co\nadsorption_energies[surface] = E_ads\n```\n3. ```python\n# Implementation C\nE_ads = E_slab_co - E_clean\nadsorption_energies[surface] = E_ads\n```\n4. ```python\n# Implementation D\nE_ads = abs(E_slab_co - E_clean - E_co_gas)\nadsorption_energies[surface] = E_ads\n```",
+        "target_scores": {
+          "```python\n# Implementation A\nE_ads = E_slab_co - E_clean - E_co_gas\nadsorption_energies[surface] = E_ads\n```": 1,
+          "```python\n# Implementation B\nE_ads = E_clean + E_co_gas - E_slab_co\nadsorption_energies[surface] = E_ads\n```": 0,
+          "```python\n# Implementation C\nE_ads = E_slab_co - E_clean\nadsorption_energies[surface] = E_ads\n```": 0,
+          "```python\n# Implementation D\nE_ads = abs(E_slab_co - E_clean - E_co_gas)\nadsorption_energies[surface] = E_ads\n```": 0
+        }
+      }
+    ],
+    "keywords": [
+      "materials_science",
+      "crystallography",
+      "surface_science",
+      "computational_chemistry",
+      "crystal_structures",
+      "adsorption",
+      "dft",
+      "mcq"
+    ],
+    "metrics": [
+      "multiple_choice_grade"
+    ],
+    "preferred_score": "multiple_choice_grade"
+  }
+]
