+++ tasks/qa/catalyst_qa/tasks_json/questions.json
@@ -470,7 +470,7 @@
       "The compiler optimization flags used for the DFT code.": 0
     },
     "explanation": "Accurate adsorption energy calculations in DFT require careful convergence of several key parameters. Slab thickness ensures that the surface properly represents the bulk, and vacuum layer size eliminates spurious interactions between periodic images. K-point sampling accurately integrates over the Brillouin zone for periodic systems. The exchange-correlation functional is the primary approximation in DFT and directly dictates the electron-electron interactions. The basis set (or plane-wave cutoff) defines the resolution of the electronic wavefunctions. The SCF convergence criterion ensures that the electronic ground state is found sufficiently accurately. Finally, the choice of pseudopotential (or all-electron treatment) defines how core electrons are handled, impacting the effective potential experienced by valence electrons. Distractors are: thermostat type (relevant for *sampling* in MD, not *accuracy* of static energy), MD simulation time (for statistical averages, not static energy), computational power (affects *speed*, not intrinsic accuracy), visualization software (post-processing), and compiler flags (affects *speed* and *numerical stability*, not fundamental physical accuracy of the model itself)."
-  }
+  },
   {
     "question": "When simulating molecular adsorption on a crystalline surface using DFT, which of the following is an important consideration or common assumption regarding the surface structure upon adsorption?",
     "options": {
@@ -521,5 +521,15 @@
       "Volcano plots demonstrate the Sabatier principle with peak activity at intermediate binding energies; the optimal d-band center corresponds to the volcano maximum": 1
     },
     "explanation": "The optimal CO2 reduction catalyst requires a balance: the d-band center position controls adsorbate binding strength, moderate CO binding satisfies the Sabatier principle (not too weak for activation, not too strong for desorption), and scaling relations between CO and CHO intermediates create fundamental trade-offs that constrain catalyst design. Breaking these scaling relations is a key challenge in rational catalyst design."
+  },
+  {
+    "question": "You're calculating CO adsorption energies on different metal surfaces from DFT calculations. Which implementation correctly calculates the adsorption energy?\n\n```python\ndef calculate_adsorption_energies(dft_results):\n    \"\"\"\n    Calculate CO adsorption energy on metal surfaces.\n    DFT results contain:\n    - 'clean_slab': Energy of clean metal slab\n    - 'co_adsorbate': Energy of CO molecule in gas phase\n    - 'slab_co': Energy of slab + adsorbed CO system\n    \"\"\"\n    adsorption_energies = {}\n    \n    for surface in ['Pt111', 'Pd111', 'Ni111']:\n        E_clean = dft_results[surface]['clean_slab']\n        E_co_gas = dft_results['CO']['co_adsorbate']\n        E_slab_co = dft_results[surface]['slab_co']\n        \n        # Different implementations:\n```",
+    "options": {
+      "```python\n# Implementation A\nE_ads = E_slab_co - E_clean - E_co_gas\nadsorption_energies[surface] = E_ads\n```": 1,
+      "```python\n# Implementation B\nE_ads = E_clean + E_co_gas - E_slab_co\nadsorption_energies[surface] = E_ads\n```": 0,
+      "```python\n# Implementation C\nE_ads = E_slab_co - E_clean\nadsorption_energies[surface] = E_ads\n```": 0,
+      "```python\n# Implementation D\nE_ads = abs(E_slab_co - E_clean - E_co_gas)\nadsorption_energies[surface] = E_ads\n```": 0
+    },
+    "explanation": "Implementation A is correct. Adsorption energy is defined as E_ads = E_slab+adsorbate - E_clean_slab - E_gas_phase_molecule. This gives negative values for favorable (exothermic) adsorption. Implementation B has the wrong sign, C is missing the gas-phase reference, and D uses absolute value which loses the thermodynamic sign information."
   }
 ]
